Finished DFA, added Tests.
